SVCPLAN-3648: register_ping_check - optionally add prefix/suffix or override domain name
The hostname set in the OS of some cluster nodes corresponds to a private IP that
is not pingable by nodes with ping checks based on exported resources.

This change allows overriding the default networking.fqdn fact that is normally
passed as an exported resource to fix this issue.

E.g., allows adding a '-mgmt' suffix to the hostname and forcing an 'internal'
domain name.